added info about `Themes` folder location
Since this is different from operating system to operating system it's worth calling out here. I spent about a few minutes scratching my head trying to figure out why I was getting theme not found errors. I assumed the Themes directory was the same on MacOS as it is on Linux.