Make theme wiki link bigger
When I first viewed the README for this repo, I completely missed the preview of all themes on the wiki. I figured it might help to make the size of the line that directs readers to the wiki to preview the themes a bit larger, which is what I did.