Branched my Theme from sleek
The creator of Sleek theme actually commented on one of my pull requests and said that I should probably branch off because we're going two seperate directions.
I plan on making Sleek theme look more boxed in with borders and so on for a more organized appeal. As well as adding some aesthetically pleasing elements. Whereas as far as I know, he wishes to keep sleek theme as clean and organized as possible