fix(Turntable/context-menu): fix FAD context menu handler invalid
Spotify v1.2.57 use React v18.3.1, and in this version of React, render API will add FAD component's context menu event listener to the `#fad-main` container element, that's why the handleFADContextMenu hander is add to this container element here. Otherwise, the handleFADContextMenu hander will be called before the FAD's hander, at this time, the config popup modal can't be get because it was created in the FAD's hander.